# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Story Adventure Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gameplay/levels/{storyId}/start:
    post:
      operationId: startLevel
      tags: [gameplay]
      summary: Start playing a story level
      parameters:
        - name: storyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSession"
  /gameplay/levels/{storyId}/complete:
    post:
      operationId: completeLevel
      tags: [gameplay]
      summary: Complete a story level and claim rewards
      parameters:
        - name: storyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LevelCompleteInput"
      responses:
        "200":
          description: Completion result with rewards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LevelCompleteResult"
  /gameplay/choices:
    post:
      operationId: makeChoice
      tags: [gameplay]
      summary: Make a story choice and get AI-generated continuation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChoiceInput"
      responses:
        "200":
          description: AI-generated story continuation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryContinuation"

  # ── REWARDS ────────────────────────────────────────────────────────────────
...
  /inventory/{itemId}/equip:
    post:
      operationId: equipItem
      tags: [inventory]
      summary: Equip an inventory item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"

  # ── BOSS BATTLES ───────────────────────────────────────────────────────────
...
  /bosses/{id}/battle:
    post:
      operationId: startBossBattle
      tags: [bosses]
      summary: Start a boss battle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Battle result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleResult"

  # ── EVENTS ─────────────────────────────────────────────────────────────────
...
  /events/{id}/join:
    post:
      operationId: joinEvent
      tags: [events]
      summary: Join a game event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventParticipation"

  # ── WITHDRAWALS ────────────────────────────────────────────────────────────
...
  /analytics/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [analytics]
      summary: Get revenue analytics over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueAnalytics"
  /analytics/users:
    get:
      operationId: getUserAnalytics
      tags: [analytics]
      summary: Get user growth and activity analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
      responses:
        "200":
          description: User analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAnalytics"
components:
  schemas:
...
    StoryChoice:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        consequence:
          type: string
      required: [id, text, consequence]
...
    GameSession:
      type: object
      properties:
        sessionId:
          type: string
        storyId:
          type: integer
        startedAt:
          type: string
        initialContent:
          type: string
        choices:
          type: array
          items:
            $ref: "#/components/schemas/StoryChoice"
      required: [sessionId, storyId, startedAt, initialContent, choices]
    LevelCompleteInput:
      type: object
      properties:
        sessionId:
          type: string
        score:
          type: integer
        timeTaken:
          type: integer
      required: [sessionId, score]
    LevelCompleteResult:
      type: object
      properties:
        coinsEarned:
          type: integer
        gemsEarned:
          type: integer
        xpEarned:
          type: integer
        newLevel:
          type: integer
        leveledUp:
          type: boolean
        newBadges:
          type: array
          items:
            type: string
        newUserBalance:
          $ref: "#/components/schemas/UserBalance"
      required: [coinsEarned, gemsEarned, xpEarned, newLevel, leveledUp, newBadges, newUserBalance]
    UserBalance:
      type: object
      properties:
        coins:
          type: integer
        gems:
          type: integer
        xp:
          type: integer
        level:
          type: integer
      required: [coins, gems, xp, level]
    ChoiceInput:
      type: object
      properties:
        sessionId:
          type: string
        choiceId:
          type: string
        storyId:
          type: integer
      required: [sessionId, choiceId, storyId]
    StoryContinuation:
      type: object
      properties:
        content:
          type: string
        choices:
          type: array
          items:
            $ref: "#/components/schemas/StoryChoice"
        isEnding:
          type: boolean
      required: [content, choices, isEnding]

    # ── REWARDS ───────────────────────────────────────────────────────────────
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [avatar, badge, powerup, cosmetic]
        imageUrl:
          type: ["string", "null"]
        isEquipped:
          type: boolean
        acquiredAt:
          type: string
      required: [id, name, description, type, isEquipped, acquiredAt]
...
    BattleResult:
      type: object
      properties:
        won:
          type: boolean
        damageDealt:
          type: integer
        coinsEarned:
          type: integer
        gemsEarned:
          type: integer
        xpEarned:
          type: integer
        newBalance:
          $ref: "#/components/schemas/UserBalance"
      required: [won, damageDealt, coinsEarned, gemsEarned, xpEarned, newBalance]

    # ── EVENTS ────────────────────────────────────────────────────────────────
...
    EventParticipation:
      type: object
      properties:
        eventId:
          type: integer
        joinedAt:
          type: string
        status:
          type: string
      required: [eventId, joinedAt, status]

    # ── WITHDRAWALS ───────────────────────────────────────────────────────────
...
    RevenueAnalytics:
      type: object
      properties:
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/RevenueDataPoint"
        totalRevenue:
          type: number
        growth:
          type: number
      required: [period, data, totalRevenue, growth]
    RevenueDataPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        premiumSales:
          type: integer
        withdrawals:
          type: number
      required: [date, revenue, premiumSales, withdrawals]
    UserAnalytics:
      type: object
      properties:
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserDataPoint"
        totalNewUsers:
          type: integer
        growthRate:
          type: number
      required: [period, data, totalNewUsers, growthRate]
    UserDataPoint:
      type: object
      properties:
        date:
          type: string
        newUsers:
          type: integer
        activeUsers:
          type: integer
      required: [date, newUsers, activeUsers]