# lib/api-zod/src/generated/types/storyContinuation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Story Adventure Game API
 * OpenAPI spec version: 0.1.0
 */
import type { StoryChoice } from './storyChoice';

export interface StoryContinuation {
  content: string;
  choices: StoryChoice[];
  isEnding: boolean;
}